implemented socket server

+++ hand_detector.py
@@ -33,5 +33,5 @@ def formalize(hand_landmarks):
                 single_hand_arr.append(temp_arr)
             single_hand_arr = numpy.asarray(single_hand_arr).flatten()
             multi_hand_arr.append(single_hand_arr)
-        return numpy.array(multi_hand_arr)
+        return numpy.array(multi_hand_arr, dtype=numpy.float64)
     return None
